+++ web/src/enterprise.scss
@@ -4,6 +4,9 @@
 
 // Enterprise styles
 @import './enterprise/campaigns/apply/ChangesetSpecList';
+@import './enterprise/campaigns/apply/ChangesetSpecListHeader';
+@import './enterprise/campaigns/apply/ChangesetSpecNode';
+@import './enterprise/campaigns/apply/VisibleChangesetSpecNode';
 @import './enterprise/campaigns/detail/CampaignActionsBar';
 @import './enterprise/campaigns/list/CampaignNode';
 @import './enterprise/campaigns/detail/changesets/ExternalChangesetNode';
