+++ web/src/enterprise/campaigns/apply/ChangesetSpecList.scss
@@ -6,15 +6,3 @@
         align-items: center;
     }
 }
-
-.justify-self-end {
-    justify-self: end;
-}
-
-.grid-row {
-    grid-column: 1 / -1;
-}
-
-.align-self-start {
-    align-self: start;
-}
