+++ web/src/enterprise/campaigns/apply/ChangesetSpecListHeader.scss
@@ -1,0 +1,6 @@
+.changeset-spec-list-header {
+    &__title {
+        // Make it full width in the current row.
+        grid-column: 1 / -1;
+    }
+}
