+++ web/src/enterprise/campaigns/apply/ChangesetSpecNode.scss
@@ -1,0 +1,15 @@
+.changeset-spec-node {
+    &__separator {
+        // Make it full width in the current row.
+        grid-column: 1 / -1;
+        border-top: 1px solid $color-light-bg-2;
+    }
+}
+
+.theme-dark {
+    .changeset-spec-node {
+        &__separator {
+            border-color: $color-light-bg-5;
+        }
+    }
+}
