+++ web/src/enterprise/campaigns/apply/HiddenChangesetSpecNode.tsx
@@ -12,22 +12,18 @@ export const HiddenChangesetSpecNode: React.FunctionComponent<HiddenChangesetSpe
     <>
         <span />
         <ChangesetSpecAction spec={node} />
-        <div className="changeset-node__content flex-fill">
-            <div className="d-flex flex-column">
-                <div className="m-0 mb-2">
-                    <h3 className="m-0 text-muted">
-                        {node.type === ChangesetSpecType.EXISTING && <>Import changeset from a private repository</>}
-                        {node.type === ChangesetSpecType.BRANCH && <>Create changeset in a private repository</>}
-                    </h3>
-                    <span className="text-danger">
-                        No action will happen on apply.{' '}
-                        <InfoCircleOutlineIcon
-                            className="icon-inline"
-                            data-tooltip="You have no permissions to access this repository."
-                        />
-                    </span>
-                </div>
-            </div>
+        <div className="d-flex flex-column">
+            <h3 className="text-muted">
+                {node.type === ChangesetSpecType.EXISTING && <>Import changeset from a private repository</>}
+                {node.type === ChangesetSpecType.BRANCH && <>Create changeset in a private repository</>}
+            </h3>
+            <span className="text-danger">
+                No action will happen on apply.{' '}
+                <InfoCircleOutlineIcon
+                    className="icon-inline"
+                    data-tooltip="You have no permissions to access this repository."
+                />
+            </span>
         </div>
         <span />
     </>
