+++ web/src/enterprise/campaigns/apply/VisibleChangesetSpecNode.scss
@@ -1,0 +1,10 @@
+.visible-changeset-spec-node {
+    &__expanded-section {
+        // Make it full width in the current row.
+        grid-column: 1 / -1;
+    }
+    &__diffstat {
+        justify-self: end;
+        align-self: start;
+    }
+}
